modified package structure. added an annotation system to automatically handle slash command argument payloads.
