--- src/main/java/delilah/discord/EventListener.java
@@ -1,26 +1,0 @@
-package delilah.discord;
-
-import delilah.commands.CommandEventHandler;
-import net.dv8tion.jda.api.events.interaction.command.CommandAutoCompleteInteractionEvent;
-import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import org.jetbrains.annotations.NotNull;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-@Component
-public class EventListener extends ListenerAdapter {
-
-    @Autowired
-    private CommandEventHandler dispatcher;
-
-    @Override
-    public void onSlashCommandInteraction(@NotNull SlashCommandInteractionEvent event) {
-        dispatcher.onSlashcommandInteractionEvent(event);
-    }
-
-    @Override
-    public void onCommandAutoCompleteInteraction(@NotNull CommandAutoCompleteInteractionEvent event) {
-        dispatcher.onCommandAutoCompleteInteraction(event);
-    }
-}
