+++ src/main/java/delilah/domain/models/dictionnary/DictionaryPermissionScope.java
@@ -1,0 +1,6 @@
+package delilah.domain.models.dictionnary;
+
+public enum DictionaryPermissionScope {
+    PUBLIC,
+    PRIVATE
+}
