+++ src/main/java/delilah/infrastructure/repositories/UserRepository.java
@@ -1,0 +1,9 @@
+package delilah.infrastructure.repositories;
+
+import delilah.domain.models.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+}
