+++ src/main/java/delilah/interf4ce/commands/payloadProcessing/annotations/ConsumesPayload.java
@@ -1,0 +1,9 @@
+package delilah.interf4ce.commands.payloadProcessing.annotations;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ConsumesPayload {
+    Class<?> type();
+}
