--- src/main/java/delilah/models/dictionnary/DictionaryPermissionScope.java
@@ -1,6 +1,0 @@
-package delilah.models.dictionnary;
-
-public enum DictionaryPermissionScope {
-    PUBLIC,
-    PRIVATE
-}
