--- src/main/java/delilah/repositories/DictionaryRepository.java
@@ -1,9 +1,0 @@
-package delilah.repositories;
-
-import delilah.models.dictionnary.Dictionary;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DictionaryRepository extends JpaRepository<Dictionary, String> {
-}
