--- src/main/java/delilah/repositories/UserRepository.java
@@ -1,9 +1,0 @@
-package delilah.repositories;
-
-import delilah.models.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, String> {
-}
